+++ packages/udp/src/index.ts
@@ -1,2 +1,2 @@
-export * from './types.js'
+export * from '@simple-discovery/core'
 export * from './UdpDiscovery.js'